+++ YnabSharp.Seeder/Step1.SetPlan/SetPlanCliCommandFactory.cs
@@ -4,16 +4,16 @@ using KitCli.Commands.Abstractions.Factories;
 using KitCli.Instructions.Abstractions;
 using KitCli.Instructions.Arguments;
 
-namespace YnabSharp.Seeder.Step1.SetBudget;
+namespace YnabSharp.Seeder.Step1.SetPlan;
 
-public class SetBudgetCliCommandFactory : ICliCommandFactory<SetBudgetCliCommand>
+public class SetPlanCliCommandFactory : ICliCommandFactory<SetPlanCliCommand>
 {
     public CliCommand Create(CliInstruction instruction, List<CliCommandArtefact> artefacts)
     {
         var countArgument = instruction
             .Arguments
-            .OfRequiredType<string>(SetBudgetCliCommand.ArgumentNames.BudgetName);
-        
-        return new SetBudgetCliCommand(countArgument.ArgumentValue);
+            .OfRequiredType<string>(SetPlanCliCommand.ArgumentNames.PlanName);
+
+        return new SetPlanCliCommand(countArgument.ArgumentValue);
     }
 }