+++ YnabSharp.Seeder/YnabRegistry.cs
@@ -2,7 +2,7 @@ using KitCli.Abstractions;
 using KitCli.Commands.Abstractions.Artefacts;
 using Microsoft.Extensions.DependencyInjection;
 using YnabSharp.Extensions;
-using YnabSharp.Seeder.Step1.SetBudget;
+using YnabSharp.Seeder.Step1.SetPlan;
 using YnabSharp.Seeder.Step2.Prepare.Step21.PrepareAccounts;
 using YnabSharp.Seeder.Step2.Prepare.Step22.PrepareTransactions;
 using YnabSharp.Seeder.Step3.Seed.Step33.SeedAccounts;
@@ -20,7 +20,7 @@ public class YnabRegistry : ICliAppBuilderRegistry
 
     private void RegisterArtefactFactories(IServiceCollection services)
     {
-        services.AddSingleton<ICliCommandArtefactFactory, BudgetCliCommandArtefactFactory>();
+        services.AddSingleton<ICliCommandArtefactFactory, PlanCliCommandArtefactFactory>();
         services.AddSingleton<ICliCommandArtefactFactory, AccountsCliCommandArtefactFactory>();
         services.AddSingleton<ICliCommandArtefactFactory, AccountIdsCliCommandArtefactFactory>();
         services.AddSingleton<ICliCommandArtefactFactory, TransactionsCliCommandArtefactFactory>();