+++ YnabSharp/Connected/ConnectedPlan.cs
@@ -1,22 +1,22 @@
 using YnabSharp.Clients;
 using YnabSharp.Mappers;
-using YnabSharp.Responses.Budgets;
+using YnabSharp.Responses.Plans;
 
 namespace YnabSharp.Connected;
 
-public class ConnectedBudget : Budget
+public class ConnectedPlan : Plan
 {
     private readonly AccountClient _accountClient;
     private readonly CategoryClient _categoryClient;
     private readonly TransactionClient _transactionClient;
     private readonly ScheduledTransactionClient _scheduledTransactionsClient;
 
-    public ConnectedBudget(
+    public ConnectedPlan(
         AccountClient accountClient,
         CategoryClient categoryClient,
         TransactionClient transactionClient,
         ScheduledTransactionClient scheduledTransactionsClient,
-        BudgetResponse budgetResponse) : base(budgetResponse)
+        PlanResponse planResponse) : base(planResponse)
     {
         _accountClient = accountClient;
         _categoryClient = categoryClient;
@@ -32,22 +32,22 @@ public class ConnectedBudget : Budget
         => _categoryClient.GetAll();
     public Task<IEnumerable<Transaction>> GetTransactions()
         => _transactionClient.GetAll();
-    public Task<Transaction> GetTransaction(string id) 
+    public Task<Transaction> GetTransaction(string id)
         => _transactionClient.Get(id);
-    public Task<IEnumerable<Transaction>> CreateTransactions(IEnumerable<Transaction> transactions) => 
+    public Task<IEnumerable<Transaction>> CreateTransactions(IEnumerable<Transaction> transactions) =>
         _transactionClient.Create(transactions);
     public Task<ConnectedAccount> CreateAccount(NewAccount newAccount)
         => _accountClient.Create(newAccount);
     public Task<ConnectedAccount> CreateAccount(Account account)
         => _accountClient.Create(account);
-    
+
     public async Task MoveAccountTransactions(ConnectedAccount fromAccount, ConnectedAccount toAccount)
     {
         var transactionsTask = fromAccount.GetTransactions();
         var scheduledTransactionsTask = fromAccount.GetScheduledTransactions();
-        
+
         await Task.WhenAll(transactionsTask, scheduledTransactionsTask);
-        
+
         var transactionsToMove = transactionsTask
             .Result
             .Where(t => t.PayeeName != AutomatedPayeeNames.StartingBalance)
@@ -60,9 +60,9 @@ public class ConnectedBudget : Budget
         var scheduledTransactionMoveTasks = scheduledTransactionsToMove
             .Select(movedScheduledTransaction =>
                 _scheduledTransactionsClient.MoveTransaction(movedScheduledTransaction));
-        
+
         await Task.WhenAll(scheduledTransactionMoveTasks);
         _  = await _transactionClient.Move(transactionsToMove);
     }
-    
+
 }