+++ YnabSharp/Responses/Plans/GetPlansResponseData.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace YnabSharp.Responses.Plans;
+
+public class GetPlansResponseData
+{
+    [JsonPropertyName("plans")]
+    public required IEnumerable<PlanResponse> Plans { get; set; }
+}