+++ YnabSharp/Responses/Plans/PlanResponse.cs
@@ -1,18 +1,18 @@
 using System.Text.Json.Serialization;
 
-namespace YnabSharp.Responses.Budgets;
+namespace YnabSharp.Responses.Plans;
 
-public class BudgetResponse
+public class PlanResponse
 {
     [JsonPropertyName("id")]
     public Guid Id { get; set; }
-    
+
     [JsonPropertyName("name")]
     public string Name { get; set; }
-    
+
     [JsonPropertyName("first_month")]
     public DateOnly FirstMonth { get; set; }
-    
+
     [JsonPropertyName("last_month")]
     public DateOnly LastMonth { get; set; }
 }